pyproject.toml
[project]
name = "spoon-agent"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "spoon-toolkits>=0.2.2",
    "spoon-ai-sdk>=0.3.1",
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    "openai>=2.8.1",
    "pydantic>=2.11.10",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.0.1",
    "rich>=14.2.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.14.5",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["spoon_agent*"]
exclude = ["tests*"]